refactor: 기능별 main 파일 분리
